requirement.txt
python2.7
pillow 
numpy
pypdf2
wand
opencv
